+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/reports/exitinfo/LatestAppExitInfoProvider.kt
@@ -11,6 +11,7 @@ import android.app.Application
 import android.app.ApplicationExitInfo
 import android.os.Build
 import androidx.annotation.RequiresApi
+import androidx.annotation.VisibleForTesting
 import io.bitdrift.capture.reports.binformat.v1.issue_reporting.ReportType
 
 /**
@@ -20,32 +21,52 @@ internal object LatestAppExitInfoProvider : ILatestAppExitInfoProvider {
     internal const val EXIT_REASON_EXCEPTION_MESSAGE =
         "LatestAppExitInfoProvider: Failed to retrieve LatestAppExitReasonResult"
 
+    /**
+     * Caching after initial fetch to avoid unnecessary IPC binder calls once value is retrieved
+     */
+    @Volatile
+    private var cachedResult: LatestAppExitReasonResult? = null
+
     @RequiresApi(Build.VERSION_CODES.R)
-    override fun get(activityManager: ActivityManager): LatestAppExitReasonResult =
-        try {
-            // a null packageName means match all packages belonging to the caller's process (UID)
-            // pid should be 0, a value of 0 means to ignore this parameter and return all matching records
-            // maxNum should be 0, this will return the list of all last exists at the time
-            val latestKnownExitReasons =
-                activityManager
-                    .getHistoricalProcessExitReasons(null, 0, 0)
+    override fun get(activityManager: ActivityManager): LatestAppExitReasonResult {
+        cachedResult?.let { return it }
 
-            val matchingProcessReason =
-                latestKnownExitReasons.firstOrNull {
-                    it.processName == Application.getProcessName()
-                }
-
-            if (matchingProcessReason == null) {
-                LatestAppExitReasonResult.None
-            } else {
-                LatestAppExitReasonResult.Valid(matchingProcessReason)
-            }
-        } catch (error: Throwable) {
-            LatestAppExitReasonResult.Error(
-                EXIT_REASON_EXCEPTION_MESSAGE,
-                error,
-            )
+        synchronized(this) {
+            cachedResult?.let { return it }
+            return getReasonAndCacheResult(activityManager)
         }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.R)
+    private fun getReasonAndCacheResult(activityManager: ActivityManager): LatestAppExitReasonResult {
+        val result =
+            try {
+                // a null packageName means match all packages belonging to the caller's process (UID)
+                // pid should be 0, a value of 0 means to ignore this parameter and return all matching records
+                // maxNum should be 0, this will return the list of all last exits at the time
+                val latestKnownExitReasons =
+                    activityManager
+                        .getHistoricalProcessExitReasons(null, 0, 0)
+
+                val matchingProcessReason =
+                    latestKnownExitReasons.firstOrNull {
+                        it.processName == Application.getProcessName()
+                    }
+
+                if (matchingProcessReason == null) {
+                    LatestAppExitReasonResult.None
+                } else {
+                    LatestAppExitReasonResult.Valid(matchingProcessReason)
+                }
+            } catch (error: Throwable) {
+                LatestAppExitReasonResult.Error(
+                    EXIT_REASON_EXCEPTION_MESSAGE,
+                    error,
+                )
+            }
+        cachedResult = result
+        return result
+    }
 
     fun mapToFatalIssueType(exitReasonType: Int): Byte? =
         when (exitReasonType) {
@@ -53,4 +74,49 @@ internal object LatestAppExitInfoProvider : ILatestAppExitInfoProvider {
             ApplicationExitInfo.REASON_CRASH_NATIVE -> ReportType.NativeCrash
             else -> null
         }
+
+    @VisibleForTesting
+    internal fun clearCache() {
+        cachedResult = null
+    }
+}
+
+/**
+ * Retrieves the latest [ApplicationExitInfo] if available.
+ */
+fun interface ILatestAppExitInfoProvider {
+    /**
+     * Returns the latest [ApplicationExitInfo] when present.
+     */
+    @RequiresApi(Build.VERSION_CODES.R)
+    fun get(activityManager: ActivityManager): LatestAppExitReasonResult
+}
+
+/**
+ * The [ApplicationExitInfo] lookup result.
+ */
+sealed class LatestAppExitReasonResult {
+    /**
+     * Returns the latest [ApplicationExitInfo] when available.
+     */
+    data class Valid(
+        /** Latest [ApplicationExitInfo] for the current process. */
+        val applicationExitInfo: ApplicationExitInfo,
+    ) : LatestAppExitReasonResult()
+
+    /**
+     * No [ApplicationExitInfo] was available.
+     * (e.g. this is expected on first app installation)
+     */
+    data object None : LatestAppExitReasonResult()
+
+    /**
+     * Returns the detailed error while trying to determine prior reasons.
+     */
+    data class Error(
+        /** Human-readable message describing the lookup failure. */
+        val message: String,
+        /** Optional throwable associated with the lookup failure. */
+        val throwable: Throwable? = null,
+    ) : LatestAppExitReasonResult()
 }