+++ platform/jvm/capture/src/test/kotlin/io/bitdrift/capture/CaptureLoggerSessionOverrideTest.kt
@@ -27,6 +27,7 @@ import io.bitdrift.capture.attributes.ClientAttributes
 import io.bitdrift.capture.providers.DateProvider
 import io.bitdrift.capture.providers.FieldValue
 import io.bitdrift.capture.providers.session.SessionStrategy
+import io.bitdrift.capture.reports.exitinfo.LatestAppExitInfoProvider
 import io.bitdrift.capture.threading.CaptureDispatchers
 import okhttp3.HttpUrl
 import org.assertj.core.api.Assertions.assertThat
@@ -65,6 +66,7 @@ class CaptureLoggerSessionOverrideTest {
 
         CaptureDispatchers.setTestExecutorService(MoreExecutors.newDirectExecutorService())
         CaptureJniLibrary.load()
+        LatestAppExitInfoProvider.clearCache()
         testServerPort = CaptureTestJniLibrary.startTestApiServer(-1)
 
         val lifecycleRegistry = LifecycleRegistry(lifecycleOwner)
@@ -76,6 +78,7 @@ class CaptureLoggerSessionOverrideTest {
     @After
     fun teardown() {
         CaptureTestJniLibrary.stopTestApiServer()
+        LatestAppExitInfoProvider.clearCache()
     }
 
     private fun contextWithAppVersion(
@@ -151,6 +154,7 @@ class CaptureLoggerSessionOverrideTest {
         // We need to shut down the logger first before starting a new one, otherwise the new one fails to initialize due to the flock
         // on the ring buffer.
         CaptureJniLibrary.shutdown(logger.loggerId)
+        LatestAppExitInfoProvider.clearCache()
 
         val newContext = contextWithAppVersion(newAppVersion, newAppVersionCode)
         val newClientAttributes = ClientAttributes(newContext, lifecycleOwner)