+++ platform/jvm/microbenchmark/src/androidTest/java/io/bitdrift/microbenchmark/LogBenchmarkTest.kt
@@ -74,26 +74,6 @@ class LogBenchmarkTest {
     }
 
     @Test
-    fun logNotMatched5Fields() {
-        startLogger()
-        val fields = buildFieldsMap(5)
-
-        benchmarkRule.measureRepeated {
-            Capture.Logger.logInfo(fields) { LOG_MESSAGE }
-        }
-    }
-
-    @Test
-    fun logNotMatched10Fields() {
-        startLogger()
-        val fields = buildFieldsMap(10)
-
-        benchmarkRule.measureRepeated {
-            Capture.Logger.logInfo(fields) { LOG_MESSAGE }
-        }
-    }
-    
-    @Test
     fun logNotMatched5000Fields() {
         startLogger()
         val fields = buildFieldsMap(5000)